camera_feed/requirements.txt
Flask==2.1.1
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
numpy==1.21.2
python-dotenv==0.19.2
